LMD-154 FIX : chat 컴포넌트 분리
