+++ client/src/pages/VideoChat/Room.js
@@ -3,20 +3,9 @@ import io from 'socket.io-client';
 import Peer from 'simple-peer';
 import styled from 'styled-components';
 import S from './style';
+import Chat from './Chat'
 import { Grid, Image, Button } from 'semantic-ui-react';
 
-// 내 채팅
-const MyRow = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 10px;
-`;
-
-// 상대방 채팅
-const PartnerRow = styled(MyRow)`
-  justify-content: flex-start;
-`;
 
 //video Container
 // margin: auto 없앰
@@ -30,12 +19,7 @@ const Container = styled.div`
   flex-direction: column;
 `;
 
-//Chat + WebRTC Container
-const AllContainer = styled.div`
-  float: right;
-`;
 
-//
 const StyledVideo = styled.video`
   height: 30%;
   width: 70%;
@@ -43,13 +27,14 @@ const StyledVideo = styled.video`
 
 const Video = props => {
   const ref = useRef();
-  console.log(props);
 
   useEffect(() => {
     props.peer.on('stream', stream => {
       ref.current.srcObject = stream;
+
     });
   }, []);
+
   // 상대방 비디오
   return <StyledVideo playsInline autoPlay ref={ref} />;
 };
@@ -119,6 +104,7 @@ const Room = props => {
         });
         //user가 나갔을 때, disconnect
         socketRef.current.on("user left", id => {
+          console.log("사용자 나감");
           const peerObj = peersRef.current.find(p => p.peerID === id);
           if(peerObj) {
             peerObj.peer.destroy();
@@ -128,9 +114,12 @@ const Room = props => {
           setPeers(peers);
         })
       });
+      console.log("useEffect 실행됨")
   }, [userVideo]); //userVideo가 업데이트 되면 useEffect 실행
+  // 빈 배열로 해놓으면 가장 처음 렌더링 될 때만 실행되고 업데이트 할 경우에는 실행 할 필요가 없는 경우엔 함수의 두 번째 파라미터로 비어있는 배열을 넣어주면 된다.
 
   function createPeer(userToSignal, callerID, stream) {
+    console.log("createPeer 생성됨");
     const peer = new Peer({
       initiator: true, //요청자 이므로 true
       trickle: false,// 이건 유투버도 뭔지 모른다함. 근데 대부분 이것을 false로 설정한다고 했음
@@ -160,7 +149,7 @@ const Room = props => {
     });
 
     peer.signal(incomingSignal);
-
+    console.log("사용자 추가됨, addpeer");
     return peer;
   }
 
@@ -186,44 +175,6 @@ const Room = props => {
     }
   };
 
-  // 위에는 video 아래는 chat
-  const [yourID, setYourID] = useState();
-  const [messages, setMessages] = useState([]);
-  const [message, setMessage] = useState('');
-
-  const socketRefChat = useRef();
-
-  useEffect(() => {
-    socketRefChat.current = io.connect('http://chat-test.live-md.com:8000');
-
-    socketRefChat.current.on('your id', id => {
-      setYourID(id);
-    });
-
-    socketRefChat.current.on('message', message => {
-      console.log('here');
-      receivedMessage(message);
-    });
-  }, []);
-
-  function receivedMessage(message) {
-    setMessages(oldMsgs => [...oldMsgs, message]);
-  }
-
-  function sendMessage(e) {
-    e.preventDefault();
-    const messageObject = {
-      body: message,
-      id: yourID,
-    };
-    setMessage('');
-    socketRefChat.current.emit('send message', messageObject);
-  }
-
-  function handleChange(e) {
-    setMessage(e.target.value);
-  }
-
   return (
     <>
       <Grid>
@@ -246,38 +197,14 @@ const Room = props => {
               </div>
             </div>
             {/* 상대방 비디오 */}
-            {peers.map((peer) => {
-              return <div className="peerVideo" key={peer.peerID}>
-                <Video key={peer.peerID} peer={peer.peer} /> {/*key값을 index가 아닌 peerID로 변경*/}
-                </div>
+            {peers.map((peer) => { {/*key값을 index가 아닌 peerID로 변경*/}
+              console.log("비디오 불림 ㅋㄷㅋㄷ");
+              return (
+                <Video  peer={peer.peer} key={peer.peerID}/>
+              );
             })}
-
-            <S.Page>
-              <S.Container>
-                {messages.map((message, index) => {
-                  if (message.id === yourID) {
-                    return (
-                      <MyRow key={index}>
-                        <S.MyMessage>{message.body}</S.MyMessage>
-                      </MyRow>
-                    );
-                  }
-                  return (
-                    <PartnerRow key={index}>
-                      <S.PartnerMessage>{message.body}</S.PartnerMessage>
-                    </PartnerRow>
-                  );
-                })}
-              </S.Container>
-              <S.Form onSubmit={sendMessage}>
-                <S.TextArea
-                  value={message}
-                  onChange={handleChange}
-                  placeholder="Say something..."
-                />
-                <S.Button>Send</S.Button>
-              </S.Form>
-            </S.Page>
+          <Chat/>
+           
           </Grid.Column>
         </Grid.Row>
       </Grid>
